@@ -1,10 +1,10 @@
 name: sliding_up_panel2
 description: A draggable Flutter widget that makes implementing a SlidingUpPanel much easier!
-version: 3.3.0+1
+version: 3.4.0+1
 homepage: https://github.com/Zotov-VD/sliding_up_panel
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.1.3 <4.0.0"
 
 dependencies:
   flutter:
